[ion/timing] Use explicit integer types

+++ ion/src/device/boot/rt0.cpp
@@ -93,7 +93,7 @@ void start() {
   abort();
 }
 
-volatile long millis_elapsed = 0;
+volatile uint32_t millis_elapsed = 0;
 
 void __attribute__((interrupt)) sysTick() {
   millis_elapsed++;
